requirements.txt
bitarray==2.9.2
colorama==0.4.6
construct==2.10.70
cramjam==2.8.3
fastparquet==0.8.1
fsspec==2023.1.0
lz4==4.3.2
msgpack==1.0.5
numpy==1.21.6
pandas==1.1.5
pybloom-live==4.0.0
python-dateutil==2.9.0.post0
python-snappy==0.7.1
pytz==2024.1
six==1.16.0
thrift==0.20.0
tqdm==4.66.4
xxhash==3.4.1
